feat: add separate exports for web3js and Kit
