+++ crates/node-litesvm/tests/nonExistentAccount.test.ts
@@ -1,6 +1,6 @@
 import { test } from "node:test";
 import assert from "node:assert/strict";
-import { LiteSVM } from "litesvm";
+import { LiteSVM } from "litesvm/kit";
 import { generateAddress } from "./util";
 
 test("non-existent account", async () => {
