+++ crates/node-litesvm/tests/oneTransfer.test.ts
@@ -1,6 +1,6 @@
 import { getTransferSolInstruction } from "@solana-program/system";
 import { generateKeyPairSigner, lamports } from "@solana/kit";
-import { LiteSVM } from "litesvm";
+import { LiteSVM } from "litesvm/kit";
 import assert from "node:assert/strict";
 import { test } from "node:test";
 import {
