+++ crates/node-litesvm/tests/usdcMint.test.ts
@@ -14,7 +14,7 @@ import {
 	lamports,
 	none,
 } from "@solana/kit";
-import { LiteSVM } from "litesvm";
+import { LiteSVM } from "litesvm/kit";
 import assert from "node:assert/strict";
 import { test } from "node:test";
 import { generateAddress, LAMPORTS_PER_SOL } from "./util";
